fix(tyr): wire Tyr to shared niuu infrastructure and fix import flow
- Add `src/niuu` to pyproject.toml wheel packages (fixes ModuleNotFoundError)
- Add Tyr ingress, Envoy sidecar, and JWT auth (same pattern as Volundr)
- Add integration_connections migration to Tyr database
- Extract InfisicalCredentialStore and API client to shared niuu module
- Move repos endpoint to /api/v1/niuu/repos (shared across services)
- Add dynamic credential store config to Tyr (replaces MemoryCredentialStore)
- Fix Linear GraphQL query types (String! for lookups, ID! for filters)
- Fix milestone progress calculation (divide Linear percentages by 100)
- Sort milestones ascending by sort_order
- Rework RepoSelector as searchable dropdown with branch selection
- Align Tyr UI colors with platform theme (use --color-brand tokens)
