+++ charts/tyr/templates/_helpers.tpl
@@ -137,4 +137,7 @@ Annotations for checksum/config - forces restart on config changes
 */}}
 {{- define "tyr.checksumAnnotations" -}}
 checksum/config: {{ include (print $.Template.BasePath "/configmap.yaml") . | sha256sum }}
+{{- if .Values.envoy.enabled }}
+checksum/envoy: {{ include (print $.Template.BasePath "/envoy-configmap.yaml") . | sha256sum }}
+{{- end }}
 {{- end }}
