+++ charts/tyr/templates/envoy-configmap.yaml
@@ -1,0 +1,134 @@
+{{- if .Values.envoy.enabled }}
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: {{ include "tyr.fullname" . }}-envoy
+  labels:
+    {{- include "tyr.labels" . | nindent 4 }}
+data:
+  envoy.yaml: |
+    static_resources:
+      listeners:
+        - name: listener_0
+          address:
+            socket_address:
+              address: 0.0.0.0
+              port_value: {{ .Values.envoy.port }}
+          filter_chains:
+            - filters:
+                - name: envoy.filters.network.http_connection_manager
+                  typed_config:
+                    "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
+                    stat_prefix: ingress_http
+                    codec_type: AUTO
+                    route_config:
+                      name: local_route
+                      virtual_hosts:
+                        - name: backend
+                          domains: ["*"]
+                          routes:
+                            - match:
+                                prefix: "/"
+                              route:
+                                cluster: local_service
+                                timeout: {{ .Values.envoy.upstreamTimeout }}
+                    http_filters:
+                      {{- if .Values.envoy.jwt.enabled }}
+                      # JWT Authentication filter
+                      - name: envoy.filters.http.jwt_authn
+                        typed_config:
+                          "@type": type.googleapis.com/envoy.extensions.filters.http.jwt_authn.v3.JwtAuthentication
+                          providers:
+                            keycloak:
+                              issuer: {{ .Values.envoy.jwt.issuer | quote }}
+                              {{- if .Values.envoy.jwt.audiences }}
+                              audiences:
+                                {{- range .Values.envoy.jwt.audiences }}
+                                - {{ . | quote }}
+                                {{- end }}
+                              {{- end }}
+                              remote_jwks:
+                                http_uri:
+                                  uri: {{ .Values.envoy.jwt.jwksUri | quote }}
+                                  cluster: keycloak
+                                  timeout: {{ .Values.envoy.jwt.jwksTimeout }}
+                                cache_duration:
+                                  seconds: {{ .Values.envoy.jwt.jwksCacheDurationSeconds }}
+                              forward: true
+                              payload_in_metadata: jwt_payload
+                              claim_to_headers:
+                                - header_name: {{ .Values.envoy.headerNames.userId }}
+                                  claim_name: sub
+                                - header_name: {{ .Values.envoy.headerNames.email }}
+                                  claim_name: email
+                                - header_name: {{ .Values.envoy.headerNames.tenant }}
+                                  claim_name: {{ .Values.envoy.jwt.tenantClaim }}
+                                - header_name: {{ .Values.envoy.headerNames.roles }}
+                                  claim_name: {{ .Values.envoy.jwt.rolesClaim }}
+                                {{- range .Values.envoy.jwt.extraClaimHeaders }}
+                                - header_name: {{ .headerName | quote }}
+                                  claim_name: {{ .claimName | quote }}
+                                {{- end }}
+                          rules:
+                            {{- range .Values.envoy.jwt.bypassPrefixes }}
+                            - match:
+                                prefix: {{ . | quote }}
+                            {{- end }}
+                            - match:
+                                prefix: "/"
+                              requires:
+                                provider_name: keycloak
+                      {{- end }}
+                      {{- with .Values.envoy.extraHttpFilters }}
+                      {{- toYaml . | nindent 22 }}
+                      {{- end }}
+                      - name: envoy.filters.http.router
+                        typed_config:
+                          "@type": type.googleapis.com/envoy.extensions.filters.http.router.v3.Router
+
+      clusters:
+        - name: local_service
+          connect_timeout: {{ .Values.envoy.connectTimeout }}
+          type: STATIC
+          lb_policy: ROUND_ROBIN
+          load_assignment:
+            cluster_name: local_service
+            endpoints:
+              - lb_endpoints:
+                  - endpoint:
+                      address:
+                        socket_address:
+                          address: 127.0.0.1
+                          port_value: {{ .Values.service.targetPort | default 8081 }}
+        {{- if .Values.envoy.jwt.enabled }}
+        - name: keycloak
+          connect_timeout: {{ .Values.envoy.connectTimeout }}
+          type: STRICT_DNS
+          lb_policy: ROUND_ROBIN
+          load_assignment:
+            cluster_name: keycloak
+            endpoints:
+              - lb_endpoints:
+                  - endpoint:
+                      address:
+                        socket_address:
+                          address: {{ .Values.envoy.jwt.keycloakHost | quote }}
+                          port_value: {{ .Values.envoy.jwt.keycloakPort }}
+          {{- if .Values.envoy.jwt.keycloakTls }}
+          transport_socket:
+            name: envoy.transport_sockets.tls
+            typed_config:
+              "@type": type.googleapis.com/envoy.extensions.transport_sockets.tls.v3.UpstreamTlsContext
+              sni: {{ .Values.envoy.jwt.keycloakHost | quote }}
+          {{- end }}
+        {{- end }}
+        {{- with .Values.envoy.extraClusters }}
+        {{- toYaml . | nindent 8 }}
+        {{- end }}
+
+    admin:
+      address:
+        socket_address:
+          address: 127.0.0.1
+          port_value: {{ .Values.envoy.adminPort }}
+{{- end }}
