+++ charts/tyr/templates/migrations-configmap.yaml
@@ -78,4 +78,26 @@ data:
     DROP TABLE IF EXISTS raids;
     DROP TABLE IF EXISTS phases;
     DROP TABLE IF EXISTS sagas;
+
+  000002_integration_connections.up.sql: |
+    -- Integration connections (shared table used by niuu adapters)
+
+    CREATE TABLE IF NOT EXISTS integration_connections (
+        id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+        user_id VARCHAR(255) NOT NULL,
+        integration_type VARCHAR(50) NOT NULL,
+        adapter VARCHAR(500) NOT NULL,
+        credential_name VARCHAR(253) NOT NULL,
+        config JSONB NOT NULL DEFAULT '{}',
+        enabled BOOLEAN NOT NULL DEFAULT true,
+        slug VARCHAR(100) NOT NULL DEFAULT '',
+        created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+        updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
+    );
+
+    CREATE INDEX IF NOT EXISTS idx_integration_connections_user
+        ON integration_connections(user_id, integration_type);
+
+  000002_integration_connections.down.sql: |
+    DROP TABLE IF EXISTS integration_connections;
 {{- end }}
