+++ charts/tyr/templates/service.yaml
@@ -13,7 +13,7 @@ spec:
   ports:
     - name: http
       port: {{ .Values.service.port | default 80 }}
-      targetPort: http
+      targetPort: {{ if .Values.envoy.enabled }}envoy{{ else }}http{{ end }}
       protocol: TCP
   selector:
     {{- include "tyr.selectorLabels" . | nindent 4 }}
