+++ pyproject.toml
@@ -63,7 +63,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/volundr", "src/tyr"]
+packages = ["src/volundr", "src/tyr", "src/niuu"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
