+++ web/src/modules/tyr/components/RepoSelector/RepoSelector.module.css
@@ -1,49 +1,98 @@
 .container {
   display: flex;
   flex-direction: column;
-  gap: var(--space-1);
+  gap: var(--space-3);
 }
 
-.item,
-.itemSelected {
+.dropdownWrapper {
+  position: relative;
+}
+
+.searchInput {
+  width: 100%;
+  padding: var(--space-2) var(--space-3);
+  font-size: var(--text-sm);
+  font-family: var(--font-sans);
+  color: var(--color-text-primary);
+  background-color: var(--color-bg-primary);
+  border: 1px solid var(--color-border);
+  border-radius: var(--radius-sm);
+  outline: none;
+  transition: border-color var(--transition-fast);
+  box-sizing: border-box;
+}
+
+.searchInput:focus {
+  border-color: var(--color-brand);
+}
+
+.searchInput::placeholder {
+  color: var(--color-text-muted);
+}
+
+.dropdown {
+  position: absolute;
+  top: calc(100% + var(--space-1));
+  left: 0;
+  right: 0;
+  max-height: 240px;
+  overflow-y: auto;
+  background-color: var(--color-bg-secondary);
+  border: 1px solid var(--color-border);
+  border-radius: var(--radius-sm);
+  z-index: 10;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+  overscroll-behavior: contain;
+}
+
+.dropdownItem {
   display: flex;
   align-items: center;
   gap: var(--space-3);
   width: 100%;
   padding: var(--space-2) var(--space-3);
   font-size: var(--text-sm);
+  font-family: var(--font-sans);
   color: var(--color-text-primary);
-  background-color: var(--color-bg-primary);
-  border: 1px solid var(--color-border);
-  border-radius: var(--radius-sm);
+  background: none;
+  border: none;
+  border-bottom: 1px solid var(--color-border-subtle);
   cursor: pointer;
   text-align: left;
-  transition:
-    background-color var(--transition-fast),
-    border-color var(--transition-fast);
+  transition: background-color var(--transition-fast);
 }
 
-.item:hover {
+.dropdownItem:last-child {
+  border-bottom: none;
+}
+
+.dropdownItem:hover {
   background-color: var(--color-bg-tertiary);
 }
 
-.itemSelected {
-  background-color: color-mix(in srgb, var(--color-accent-indigo) 10%, transparent);
-  border-color: color-mix(in srgb, var(--color-accent-indigo) 40%, transparent);
+.noResults {
+  padding: var(--space-3);
+  text-align: center;
+  color: var(--color-text-muted);
+  font-size: var(--text-sm);
 }
 
-.itemSelected:hover {
-  background-color: color-mix(in srgb, var(--color-accent-indigo) 15%, transparent);
+.selectedList {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-1);
 }
 
-.checkbox {
-  font-size: var(--text-base);
-  color: var(--color-text-secondary);
-  flex-shrink: 0;
-}
-
-.itemSelected .checkbox {
-  color: var(--color-accent-indigo);
+.selectedItem {
+  display: flex;
+  align-items: center;
+  gap: var(--space-3);
+  padding: var(--space-2) var(--space-3);
+  font-size: var(--text-sm);
+  color: var(--color-text-primary);
+  background-color: color-mix(in srgb, var(--color-brand) 10%, transparent);
+  border: 1px solid color-mix(in srgb, var(--color-brand) 20%, transparent);
+  border-radius: var(--radius-sm);
 }
 
 .repoName {
@@ -62,8 +111,8 @@
   font-weight: 500;
   text-transform: uppercase;
   letter-spacing: 0.05em;
-  color: var(--color-accent-cyan);
-  background-color: color-mix(in srgb, var(--color-accent-cyan) 15%, transparent);
+  color: var(--color-text-secondary);
+  background-color: var(--color-bg-elevated);
   border-radius: var(--radius-full);
   flex-shrink: 0;
 }
@@ -73,6 +122,55 @@
   font-size: var(--text-xs);
   color: var(--color-text-muted);
   flex-shrink: 0;
+}
+
+.branchSelect {
+  padding: var(--space-1) var(--space-2);
+  font-family: var(--font-mono);
+  font-size: var(--text-xs);
+  color: var(--color-text-primary);
+  background-color: var(--color-bg-primary);
+  border: 1px solid var(--color-border);
+  border-radius: var(--radius-sm);
+  cursor: pointer;
+  flex-shrink: 0;
+  max-width: 180px;
+  outline: none;
+  transition: border-color var(--transition-fast);
+}
+
+.branchSelect:focus {
+  border-color: var(--color-brand);
+}
+
+.branchLabel {
+  font-family: var(--font-mono);
+  font-size: var(--text-xs);
+  color: var(--color-text-muted);
+  flex-shrink: 0;
+}
+
+.removeButton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 20px;
+  height: 20px;
+  font-size: var(--text-xs);
+  color: var(--color-text-muted);
+  background: none;
+  border: none;
+  border-radius: var(--radius-full);
+  cursor: pointer;
+  flex-shrink: 0;
+  transition:
+    color var(--transition-fast),
+    background-color var(--transition-fast);
+}
+
+.removeButton:hover {
+  color: var(--color-accent-red);
+  background-color: color-mix(in srgb, var(--color-accent-red) 15%, transparent);
 }
 
 .empty {
