+++ web/src/modules/tyr/pages/ImportView/ImportView.module.css
@@ -132,16 +132,16 @@
   padding: var(--space-2) var(--space-4);
   font-size: var(--text-sm);
   font-weight: 500;
-  color: var(--color-accent-indigo);
-  background-color: color-mix(in srgb, var(--color-accent-indigo) 15%, transparent);
-  border: 1px solid color-mix(in srgb, var(--color-accent-indigo) 30%, transparent);
+  color: var(--color-brand);
+  background-color: color-mix(in srgb, var(--color-brand) 15%, transparent);
+  border: 1px solid color-mix(in srgb, var(--color-brand) 30%, transparent);
   border-radius: var(--radius-sm);
   cursor: pointer;
   transition: background-color var(--transition-fast);
 }
 
 .importButton:hover:not(:disabled) {
-  background-color: color-mix(in srgb, var(--color-accent-indigo) 25%, transparent);
+  background-color: color-mix(in srgb, var(--color-brand) 25%, transparent);
 }
 
 .importButton:disabled {
