+++ web/src/modules/tyr/ports/tracker.port.ts
@@ -1,4 +1,4 @@
-import type { TrackerProject, TrackerMilestone, TrackerIssue, RepoInfo } from '../models';
+import type { TrackerProject, TrackerMilestone, TrackerIssue } from '../models';
 import type { Saga } from '../models';
 
 export interface ITrackerBrowserService {
@@ -6,6 +6,5 @@ export interface ITrackerBrowserService {
   getProject(projectId: string): Promise<TrackerProject>;
   listMilestones(projectId: string): Promise<TrackerMilestone[]>;
   listIssues(projectId: string, milestoneId?: string): Promise<TrackerIssue[]>;
-  listRepos(): Promise<RepoInfo[]>;
   importProject(projectId: string, repos: string[]): Promise<Saga>;
 }
